Test Minigame functional complete
Test minigame works in 2d scene. Timer increases from 0-10. A target range is set from 3-8. When in range for 1 second, timer changes to yellow. If space bar is pressed when in range, timer stops and text changes green. A success message displays.
Code is saved in timing_minigame_test script